requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
astroid==3.1.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
django-debug-toolbar==4.3.0
exceptiongroup==1.2.0
executing==2.0.1
flake8==7.0.0
fonttools==4.51.0
idna==3.7
ipython==8.23.0
isort==5.13.2
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mccabe==0.7.0
netaddr==1.2.1
numpy==1.26.4
packaging==24.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
Pygments==2.17.2
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
qrcode==7.4.2
requests==2.31.0
segno==1.6.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==5.5.0
tomli==2.0.1
tomlkit==0.12.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
